### 2023-09-15 #5849 @editorialbot
[PRE REVIEW]: GRFolres: A code for modified gravity simulations in strong gravity
**Submitting author:** @llibert94 (<a href="http://orcid.org/0000-0002-3812-8523">Llibert Aresté Saló</a>)
**Repository:** https://github.com/GRChombo/GRFolres
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
