### 2023-10-27 #5849 @eloisabentivegna
@JLRipley314, thanks for letting me know.

@mzilhao, this would work fine, thanks! JOSS' philosophy, as well as detailed reviewing instructions, are discussed in https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. In a nutshell, a review entails downloading and using the software, as well as checking that the code base and the accompanying paper satisfy some requirements. The review is documented in a Github issue like this one, and all requirements are provided in a checklist for your convenience.

As soon as I have identified a second reviewer, I will open the review issue, where you will receive more instructions. I will remain available for any support or clarification throughout. :)
