### 2023-12-11 #5849 @eugenealim
Hi Eloisa,

Good to hear from you! As GRFolres is heavily built upon GRChombo which I
co-founded as a code and  whose authors are people I know and in some cases
are writing letters of recommendations for, I don't think I can be an
impartial reviewer! But I can recommend Justin Ripley ***@***.***)
as an alternate!

Regards,
Eugene

On Mon, Dec 11, 2023 at 8:40 AM Eloisa Bentivegna ***@***.***>
wrote:

> 👋 @eugenealim <https://github.com/eugenealim>, are you available to
> review this submission?
>
