### 2023-12-11 #5849 @eugenealim
Hi Eloisa,

Ah. How about Miguel Bezares (
https://www.nottingham.ac.uk/Mathematics/People/miguel.bezaresfigueroa)?

Regards,
Eugene

On Mon, Dec 11, 2023 at 11:51 AM Eloisa Bentivegna ***@***.***>
wrote:

> Thanks, @eugenealim <https://github.com/eugenealim>! We've asked Justin
> but he's not available. Any further suggestion would be very welcome.
>
