### 2024-02-18 #6369 @editorialbot
[REVIEW]: GRFolres: A code for modified gravity simulations in strong gravity
**Submitting author:** @llibert94 (<a href="http://orcid.org/0000-0002-3812-8523">Llibert Aresté Saló</a>)
**Repository:** https://github.com/GRChombo/GRFolres
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
