### 2024-02-28 #6369 @eloisabentivegna
@llibert94, could be begin by fixing the missing DOI reported above? The editorial bot has suggestions for what the DOI should be, which you can check and insert in your biblio file.
