### 2024-02-29 #6369 @llibert94
> @llibert94, could be begin by fixing the missing DOI reported above? The editorial bot has suggestions for what the DOI should be, which you can check and insert in your biblio file.

I have now updated the references
