### 2024-03-22 #6369 @mzilhao
> Hello, everyone! @mzilhao @chloer3, I don't see much activity on the checklists. Is there any blocker I can help with?

hi @eloisabentivegna, sorry, I'm extremely busy this semester with classes. I've now ticked some of the boxes on the checklist, I'll try to go over the remaining ones over the next few weeks.
