### 2024-03-26 #6369 @chloer3
> Hello, everyone! @mzilhao @chloer3, I don't see much activity on the checklists. Is there any blocker I can help with?

Hi @eloisabentivegna I apologize for the delay, I will be focusing on getting through the checklist this week. I will reach out if I run into issues, and thank you for reaching out
