### 2024-04-02 #6369 @llibert94
Hi @mzilhao !
Yes that's right, we do not have a doxygen documentation for GRFolres. We do have one in GRChombo but we have realised that almost nobody uses it. Furthermore, the extensions that we have released of GRChombo (such as GRDzhadzha, which has been recently published in JOSS) also do not have a doxygen documentation.
With regards to the automated tests, it is explained in the wiki (https://github.com/GRTLCollaboration/GRFolres/wiki/Compiling-the-code) how to compile and run the tests that we have included in GRFolres. Moreover, these tests together with the rest from GRChombo are checked every time that there is a push or a pull request.
Regarding the extra citation, I'll add it. Thanks for noticing it!
