### 2024-04-02 #6369 @chloer3
Hi everyone, I have compiled the code and will begin running the tests. I will be at the APS Meeting thsi week, so I won't have a ton of time, but I will continue to work on getting the tests running. Quick question, should I run any of the tests in GRChombo before running tests from GRFolres? Or should I be okay to proceed once I've compiled Chombo, GRChombo, and GRFolres?
