### 2024-04-02 #6369 @llibert94
Hi @chloer3 ! The tests in GRFolres are independent from the ones in GRChombo. After compiling Chombo, you could compile and run the tests in GRFolres (no need to compile anything or run any tests in GRChombo, just need to have downloaded it and export GRChombo_Source).
