### 2024-04-11 #6369 @chloer3
Hi everyone, I have submitted a test and am waiting for the results. In the meantime, I have a quick question. For the Performance bullet under Functionality on the checklist, could someone please specify what exactly I should be looking for to check this off? Thanks!
