### 2024-04-11 #6369 @eloisabentivegna
Hi @chloer3, the performance box refers to any performance claims the paper makes. Examples include whether the package provides a speedup with respect to benchmarks, whether it efficiently scales up (if parallel), or simply any reported runtime/memory footprint/etc. If the paper contains no such claims, you can tick the box off right away.
