### 2024-04-15 #6369 @chloer3
Hi everyone, as an update from my side, I ran the BinaryBH4dST example and am wrapping up the functionality section of the checklist. I only ran it for a short amount of time to confirm it is working, but I just wanted to check if there are specific results I should recover. The paper does not seem to mention any specifically.

@llibert94 One question I do want to ask is what are reasonable/expected values for the constraint norms? It seems like the results in the paper by D. Doneva et al (2307.06474) report values no larger than ~10^-2 for the L^2 norm of the Hamiltonian constraint. What about the momentum constraint?
