### 2024-04-16 #6369 @llibert94
Hi @chloer3 ! Yes exactly we are good as long as the norm of the Hamiltonian constraint is around 10^(-2), which can be improved with a better resolution (note that there might be higher violation of the constraint inside the Apparent Horizon because of the excision of the Gauss-Bonnet terms there).
As for the momentum constraints, they are usually one order of magnitude higher than the Hamiltonian, since the damping is more effective on the Hamiltonian than the momentum constraints. Also in general it's more important to satisfy the Hamiltonian than the momentum constraints.
Hope this helps. Let me know if you have any other question.
