### 2024-04-23 #6369 @chloer3
Hi @eloisabentivegna I am also done with my checklist. My only comment would be in regards to running tests, I think it would be helpful to include some sort of benchmark, i.e. specifying reasonable constraint results, or providing some sort of 1D plot the user should be able to recreate after running a short simulation with the provided param.txt file.

Please let me know if there anything else you need from me, thanks!
