### 2024-05-01 #6369 @eloisabentivegna
@chloer3, thanks for the suggestion. @llibert94, would it be possible to include a reference test result along with the test infrastructure? I think this is a very valid point.

@mzilhao, thanks for the recommendation! Nothing else is required of you.
