### 2024-05-07 #6369 @llibert94
Hi everyone! It's updated now! I've changed the KerrBH4dST example so that one can compare some results as explained in the section "Existing examples" from the wiki
