### 2024-05-16 #6369 @eloisabentivegna
@llibert94, could you take care of the missing DOIs above? The editorial bot is not finding suggestions, but DOIs do exist for some of the entries in the MISSING DOIs list.
