### 2024-05-17 #6369 @llibert94
Hi @eloisabentivegna ! I indeed found the DOI for some of those references and updated them
