### 2024-05-21 #6369 @eloisabentivegna
@llibert94, please include all the DOIs you can find -- this is critical for ensuring all work can be credited. I believe the first publication in the "MISSING DOIs" section has DOI 10.1017/CBO9781139172189. The second is probably 10.48550/arXiv.2112.06861 (although I see that there is also a published version of this, could you make sure your citation points to the latter if appropriate?).

Please ensure that if a DOI exist, it is included.
