### 2024-06-04 #6369 @eloisabentivegna
Hi @llibert94, I've just opened an issue with a few tiny modifications to the main text. Everything else looks good to me.

Once you've taken care of the text, please proceed with the post-review checklist below.
