### 2024-06-05 #6369 @llibert94
Hi @eloisabentivegna ! I have just implemented the changes to the paper in the main branch of the code and closed the issue!
Regarding the post-review checklist above, I do not understand at all what I have to do. Authors and affiliations (with the ORCIDs) are correct. Then I guess that the 2nd point is also done after implementing these last changes of the paper. I'm not sure though what I'm meant to do regarding the 3rd point
