### 2024-06-05 #6369 @eloisabentivegna
Hi @llibert94, for the second point we expect you to created a **tagged** release of the code, which is then quoted in the paper, indicating which version of your code the paper and associated reviews are based on.

For the third version, you should upload this version of your code onto Zenodo or another archival service providing a DOI, and post that DOI here.

Please let me know if you need more details on how to go about either of those tasks. Alternatively, you can look at other published JOSS submissions for examples of tagging and archiving your code: https://github.com/openjournals/joss-reviews/issues?q=is%3Aclosed+label%3Apublished+
