### 2024-06-06 #6369 @llibert94
Hi @eloisabentivegna ! Thanks I just completed everything in the checklist now!
The version is **v0.1.0**
The doi of the release on Zenodo is [10.5281/zenodo.11214736](https://zenodo.org/doi/10.5281/zenodo.11214736)
